+++ src/main/java/org/dbtools/android/domain/RxAndroidBaseManager.java
@@ -7,6 +7,7 @@ import org.dbtools.android.domain.database.DatabaseWrapper;
 import org.dbtools.android.domain.dbtype.DatabaseValue;
 import org.dbtools.android.domain.dbtype.DatabaseValueUtil;
 import rx.Observable;
+import rx.Subscriber;
 import rx.functions.Func0;
 
 import javax.annotation.Nonnull;
@@ -127,91 +128,51 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
     }
 
     @Nonnull
-    public Observable<List<T>> findAllRx() {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return findAllBySelection(null, null, null);
-            }
-        });
+    public Observable<T> findAllRx() {
+        return findAllBySelectionRx(null, null, null);
     }
 
     @Nonnull
-    public Observable<List<T>> findAllRx(@Nonnull final String databaseName) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return findAllBySelection(databaseName, null, null, null);
-            }
-        });
+    public Observable<T> findAllRx(@Nonnull final String databaseName) {
+        return findAllBySelectionRx(databaseName, null, null, null);
     }
 
     @Nonnull
-    public Observable<List<T>> findAllOrderByRx(@Nullable final String orderBy) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return findAllBySelection(null, null, orderBy);
-            }
-        });
+    public Observable<T> findAllOrderByRx(@Nullable final String orderBy) {
+        return findAllBySelectionRx(null, null, orderBy);
     }
 
     @Nonnull
-    public Observable<List<T>> findAllOrderByRx(@Nonnull final String databaseName, @Nullable final String orderBy) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return findAllBySelection(databaseName, null, null, orderBy);
-            }
-        });
+    public Observable<T> findAllOrderByRx(@Nonnull final String databaseName, @Nullable final String orderBy) {
+        return findAllBySelectionRx(databaseName, null, null, orderBy);
     }
 
     @Nonnull
-    public Observable<List<T>> findAllBySelectionRx(@Nullable final String selection, @Nonnull final String[] selectionArgs) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return findAllBySelection(selection, selectionArgs, null);
-            }
-        });
+    public Observable<T> findAllBySelectionRx(@Nullable final String selection, @Nonnull final String[] selectionArgs) {
+        return findAllBySelectionRx(selection, selectionArgs, null);
     }
 
     @Nonnull
-    public Observable<List<T>> findAllBySelectionRx(@Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                Cursor cursor = findCursorBySelection(selection, selectionArgs, orderBy);
-                return getAllItemsFromCursor(cursor);
-            }
-        });
+    public Observable<T> findAllBySelectionRx(@Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
+        Cursor cursor = findCursorBySelection(selection, selectionArgs, orderBy);
+        return getAllItemsFromCursorRx(cursor);
     }
 
     @Nonnull
-    public Observable<List<T>> findAllBySelectionRx(@Nonnull final String databaseName, @Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                Cursor cursor = findCursorBySelection(databaseName, selection, selectionArgs, orderBy);
-                return getAllItemsFromCursor(cursor);
-            }
-        });
+    public Observable<T> findAllBySelectionRx(@Nonnull final String databaseName, @Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
+        Cursor cursor = findCursorBySelection(databaseName, selection, selectionArgs, orderBy);
+        return getAllItemsFromCursorRx(cursor);
     }
 
     /**
      * Populate of List from a rawQuery.  The raw query must contain all of the columns names for the object
      *
      * @param rawQuery Custom query
-     * @return List of object T
+     * @return Observable<T>
      */
     @Nonnull
-    public Observable<List<T>> findAllByRawQueryRx(@Nonnull final String rawQuery) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return getAllItemsFromCursor(findCursorByRawQuery(getDatabaseName(), rawQuery, null));
-            }
-        });
+    public Observable<T> findAllByRawQueryRx(@Nonnull final String rawQuery) {
+        return getAllItemsFromCursorRx(findCursorByRawQuery(getDatabaseName(), rawQuery, null));
     }
 
     /**
@@ -219,16 +180,11 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      *
      * @param rawQuery      Custom query
      * @param selectionArgs query arguments
-     * @return List of object T
+     * @return Observable<T>
      */
     @Nonnull
-    public Observable<List<T>> findAllByRawQueryRx(@Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return getAllItemsFromCursor(findCursorByRawQuery(getDatabaseName(), rawQuery, selectionArgs));
-            }
-        });
+    public Observable<T> findAllByRawQueryRx(@Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
+        return getAllItemsFromCursorRx(findCursorByRawQuery(getDatabaseName(), rawQuery, selectionArgs));
     }
 
     /**
@@ -237,24 +193,29 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @param databaseName  Name of database
      * @param rawQuery      Custom query
      * @param selectionArgs query arguments
-     * @return List of object T
+     * @return Observable<T>
      */
     @Nonnull
-    public Observable<List<T>> findAllByRawQueryRx(@Nonnull final String databaseName, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return getAllItemsFromCursor(findCursorByRawQuery(databaseName, rawQuery, selectionArgs));
-            }
-        });
+    public Observable<T> findAllByRawQueryRx(@Nonnull final String databaseName, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
+        return getAllItemsFromCursorRx(findCursorByRawQuery(databaseName, rawQuery, selectionArgs));
     }
 
     @Nonnull
-    public Observable<List<T>> getAllItemsFromCursorRx(@Nullable final Cursor cursor) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
+    public Observable<T> getAllItemsFromCursorRx(@Nullable final Cursor cursor) {
+        return Observable.create(new Observable.OnSubscribe<T>() {
             @Override
-            public List<T> call() {
-                return getAllItemsFromCursor(cursor);
+            public void call(Subscriber<? super T> subscriber) {
+                if (cursor != null) {
+                    if (cursor.moveToFirst()) {
+                        do {
+                            T record = newRecord();
+                            record.setContent(cursor);
+                            subscriber.onNext(record);
+                        } while (cursor.moveToNext());
+                    }
+                    cursor.close();
+                    subscriber.onCompleted();
+                }
             }
         });
     }
@@ -332,33 +293,26 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
     }
 
     @Nonnull
-    public Observable<List<T>> findAllByRowIdsRx(final long[] rowIds) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return findAllByRowIds(rowIds, null);
-            }
-        });
+    public Observable<T> findAllByRowIdsRx(final long[] rowIds) {
+        return findAllByRowIdsRx(rowIds, null);
     }
 
     @Nonnull
-    public Observable<List<T>> findAllByRowIdsRx(final long[] rowIds, @Nullable final String orderBy) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return findAllByRowIds(getDatabaseName(), rowIds, orderBy);
-            }
-        });
+    public Observable<T> findAllByRowIdsRx(final long[] rowIds, @Nullable final String orderBy) {
+        return findAllByRowIdsRx(getDatabaseName(), rowIds, orderBy);
     }
 
     @Nonnull
-    public Observable<List<T>> findAllByRowIdsRx(@Nonnull final String databaseName, final long[] rowIds, @Nullable final String orderBy) {
-        return DBToolsRxUtil.just(new Func0<List<T>>() {
-            @Override
-            public List<T> call() {
-                return findAllByRowIds(databaseName, rowIds, orderBy);
+    public Observable<T> findAllByRowIdsRx(@Nonnull final String databaseName, final long[] rowIds, @Nullable final String orderBy) {
+        StringBuilder sb = new StringBuilder();
+        for (long rowId : rowIds) {
+            if (sb.length() > 0) {
+                sb.append(" OR ");
             }
-        });
+            sb.append(getPrimaryKey()).append(" = ").append(rowId);
+        }
+
+        return findAllBySelectionRx(databaseName, sb.toString(), null, orderBy);
     }
 
     public Observable<Long> findCountRx() {
@@ -446,7 +400,6 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @return total count
      */
     public Observable<Long> findCountByRawQueryRx(@Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
-
         return DBToolsRxUtil.just(new Func0<Long>() {
             @Override
             public Long call() {
@@ -484,12 +437,7 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @return query results value or defaultValue if no data was returned
      */
     public <I> Observable<I> findValueByRawQueryRx(@Nonnull final Class<I> valueType, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs, final I defaultValue) {
-        return DBToolsRxUtil.just(new Func0<I>() {
-            @Override
-            public I call() {
-                return findValueByRawQuery(getDatabaseName(), valueType, rawQuery, selectionArgs, defaultValue);
-            }
-        });
+        return findValueByRawQueryRx(getDatabaseName(), valueType, rawQuery, selectionArgs, defaultValue);
     }
 
     /**
@@ -503,12 +451,7 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @return query results value or defaultValue if no data was returned
      */
     public <I> Observable<I> findValueByRawQueryRx(@Nonnull final String databaseName, @Nonnull final Class<I> valueType, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs, final I defaultValue) {
-        return DBToolsRxUtil.just(new Func0<I>() {
-            @Override
-            public I call() {
-                return findValueByRawQuery(getReadableDatabase(databaseName), valueType, rawQuery, selectionArgs, defaultValue);
-            }
-        });
+        return findValueByRawQueryRx(getReadableDatabase(databaseName), valueType, rawQuery, selectionArgs, defaultValue);
     }
 
     /**
@@ -522,10 +465,23 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @return query results value or defaultValue if no data was returned
      */
     public static <I> Observable<I> findValueByRawQueryRx(@Nonnull final DatabaseWrapper database, @Nonnull final Class<I> valueType, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs, final I defaultValue) {
-        return DBToolsRxUtil.just(new Func0<I>() {
+        return Observable.create(new Observable.OnSubscribe<I>() {
             @Override
-            public I call() {
-                return findValueByRawQuery(database, valueType, rawQuery, selectionArgs, defaultValue);
+            public void call(Subscriber<? super I> subscriber) {
+
+                DatabaseValue<I> databaseValue = getDatabaseValue(valueType);
+                I value = defaultValue;
+
+                Cursor c = database.rawQuery(rawQuery, selectionArgs);
+                if (c != null) {
+                    if (c.moveToFirst()) {
+                        value = databaseValue.getColumnValue(c, 0, defaultValue);
+                    }
+                    c.close();
+                }
+
+                subscriber.onNext(value);
+                subscriber.onCompleted();
             }
         });
     }
@@ -541,12 +497,7 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @return query results value or defaultValue if no data was returned
      */
     public <I> Observable<I> findValueBySelectionRx(@Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, final I defaultValue) {
-        return DBToolsRxUtil.just(new Func0<I>() {
-            @Override
-            public I call() {
-                return findValueBySelection(getDatabaseName(), valueType, column, selection, selectionArgs, defaultValue);
-            }
-        });
+        return findValueBySelectionRx(getDatabaseName(), valueType, column, selection, selectionArgs, defaultValue);
     }
 
     /**
@@ -561,12 +512,7 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @return query results value or defaultValue if no data was returned
      */
     public <I> Observable<I> findValueBySelectionRx(@Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, final String orderBy, final I defaultValue) {
-        return DBToolsRxUtil.just(new Func0<I>() {
-            @Override
-            public I call() {
-                return findValueBySelection(getDatabaseName(), valueType, column, selection, selectionArgs, orderBy, defaultValue);
-            }
-        });
+        return findValueBySelectionRx(getDatabaseName(), valueType, column, selection, selectionArgs, orderBy, defaultValue);
     }
 
     /**
@@ -581,12 +527,7 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @return query results value or defaultValue if no data was returned
      */
     public <I> Observable<I> findValueBySelectionRx(@Nonnull final String databaseName, @Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, final I defaultValue) {
-        return DBToolsRxUtil.just(new Func0<I>() {
-            @Override
-            public I call() {
-                return findValueBySelection(databaseName, valueType, column, selection, selectionArgs, null, defaultValue);
-            }
-        });
+        return findValueBySelectionRx(databaseName, valueType, column, selection, selectionArgs, null, defaultValue);
     }
 
     /**
@@ -602,12 +543,7 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @return query results value or defaultValue if no data was returned
      */
     public <I> Observable<I> findValueBySelectionRx(@Nonnull final String databaseName, @Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, final String orderBy, final I defaultValue) {
-        return DBToolsRxUtil.just(new Func0<I>() {
-            @Override
-            public I call() {
-                return findValueBySelection(getReadableDatabase(databaseName), getTableName(), valueType, column, selection, selectionArgs, orderBy, defaultValue);
-            }
-        });
+        return findValueBySelectionRx(getReadableDatabase(databaseName), getTableName(), valueType, column, selection, selectionArgs, orderBy, defaultValue);
     }
 
     /**
@@ -624,10 +560,22 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @return query results value or defaultValue if no data was returned
      */
     public static <I> Observable<I> findValueBySelectionRx(@Nonnull final DatabaseWrapper database, @Nonnull final String tableName, @Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, final String orderBy, final I defaultValue) {
-        return DBToolsRxUtil.just(new Func0<I>() {
+        return Observable.create(new Observable.OnSubscribe<I>() {
             @Override
-            public I call() {
-                return findValueBySelection(database, tableName, valueType, column, selection, selectionArgs, orderBy, defaultValue);
+            public void call(Subscriber<? super I> subscriber) {
+                DatabaseValue<I> databaseValue = getDatabaseValue(valueType);
+                I value = defaultValue;
+
+                Cursor c = database.query(false, tableName, new String[]{column}, selection, selectionArgs, null, null, orderBy, "1");
+                if (c != null) {
+                    if (c.moveToFirst()) {
+                        value = databaseValue.getColumnValue(c, 0, defaultValue);
+                    }
+                    c.close();
+                }
+
+                subscriber.onNext(value);
+                subscriber.onCompleted();
             }
         });
     }
@@ -638,35 +586,11 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @param valueType     Type to be used when getting data from database and what type is used on return (Integer.class, Boolean.class, etc)
      * @param rawQuery      Query contain first column contains value
      * @param selectionArgs Query parameters
-     * @return query results List or empty List returned
+     * @return Observable of all values
      */
     @Nonnull
-    public <I> Observable<List<I>> findAllValuesByRawQueryRx(@Nonnull final Class<I> valueType, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
-        return DBToolsRxUtil.just(new Func0<List<I>>() {
-            @Override
-            public List<I> call() {
-                return findAllValuesByRawQuery(getDatabaseName(), valueType, rawQuery, selectionArgs);
-            }
-        });
-    }
-
-    /**
-     * Return a list of all of the first column values as a List for given rawQuery and selectionArgs.
-     *
-     * @param databaseName  Name of database to query
-     * @param valueType     Type to be used when getting data from database and what type is used on return (Integer.class, Boolean.class, etc)
-     * @param rawQuery      Query contain first column contains value
-     * @param selectionArgs Query parameters
-     * @return query results List or empty List returned
-     */
-    @Nonnull
-    public <I> Observable<List<I>> findAllValuesByRawQueryRx(@Nonnull final String databaseName, @Nonnull final Class<I> valueType, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
-        return DBToolsRxUtil.just(new Func0<List<I>>() {
-            @Override
-            public List<I> call() {
-                return findAllValuesByRawQuery(databaseName, valueType, 0, rawQuery, selectionArgs);
-            }
-        });
+    public <I> Observable<I> findAllValuesByRawQueryRx(@Nonnull final Class<I> valueType, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
+        return findAllValuesByRawQueryRx(getDatabaseName(), valueType, rawQuery, selectionArgs);
     }
 
     /**
@@ -676,17 +600,25 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @param valueType     Type to be used when getting data from database and what type is used on return (Integer.class, Boolean.class, etc)
      * @param rawQuery      Query contain first column contains value
      * @param selectionArgs Query parameters
-     * @return query results List or empty List returned
+     * @return Observable of all values
      */
     @Nonnull
-    public <I> Observable<List<I>> findAllValuesByRawQueryRx(@Nonnull final String databaseName, @Nonnull final Class<I> valueType, final int columnIndex, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
-        return DBToolsRxUtil.just(new Func0<List<I>>() {
-            @Override
-            public List<I> call() {
-                return findAllValuesByRawQuery(getReadableDatabase(databaseName), valueType, columnIndex, rawQuery, selectionArgs);
-            }
-        });
+    public <I> Observable<I> findAllValuesByRawQueryRx(@Nonnull final String databaseName, @Nonnull final Class<I> valueType, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
+        return findAllValuesByRawQueryRx(databaseName, valueType, 0, rawQuery, selectionArgs);
+    }
 
+    /**
+     * Return a list of all of the first column values as a List for given rawQuery and selectionArgs.
+     *
+     * @param databaseName  Name of database to query
+     * @param valueType     Type to be used when getting data from database and what type is used on return (Integer.class, Boolean.class, etc)
+     * @param rawQuery      Query contain first column contains value
+     * @param selectionArgs Query parameters
+     * @return Observable of all values
+     */
+    @Nonnull
+    public <I> Observable<I> findAllValuesByRawQueryRx(@Nonnull final String databaseName, @Nonnull final Class<I> valueType, final int columnIndex, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
+        return findAllValuesByRawQueryRx(getReadableDatabase(databaseName), valueType, columnIndex, rawQuery, selectionArgs);
     }
 
     /**
@@ -696,13 +628,23 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @param valueType     Type to be used when getting data from database and what type is used on return (Integer.class, Boolean.class, etc)
      * @param rawQuery      Query contain first column contains value
      * @param selectionArgs Query parameters
-     * @return query results List or empty List returned
+     * @return Observable of all values
      */
-    public static <I> Observable<List<I>> findAllValuesByRawQueryRx(@Nonnull final DatabaseWrapper database, @Nonnull final Class<I> valueType, final int columnIndex, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
-        return DBToolsRxUtil.just(new Func0<List<I>>() {
+    public static <I> Observable<I> findAllValuesByRawQueryRx(@Nonnull final DatabaseWrapper database, @Nonnull final Class<I> valueType, final int columnIndex, @Nonnull final String rawQuery, @Nullable final String[] selectionArgs) {
+        return Observable.create(new Observable.OnSubscribe<I>() {
             @Override
-            public List<I> call() {
-                return findAllValuesByRawQuery(database, valueType, columnIndex, rawQuery, selectionArgs);
+            public void call(Subscriber<? super I> subscriber) {
+                DatabaseValue<I> databaseValue = getDatabaseValue(valueType);
+
+                Cursor cursor = database.rawQuery(rawQuery, selectionArgs);
+                if (cursor != null) {
+                    if (cursor.moveToFirst()) {
+                        do {
+                            subscriber.onNext(databaseValue.getColumnValue(cursor, columnIndex, null));
+                        } while (cursor.moveToNext());
+                    }
+                    cursor.close();
+                }
             }
         });
     }
@@ -714,16 +656,11 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @param column        Column which contains value
      * @param selection     Query selection
      * @param selectionArgs Query parameters
-     * @return query results List or empty List returned
+     * @return Observable of all values
      */
     @Nonnull
-    public <I> Observable<List<I>> findAllValuesBySelectionRx(@Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs) {
-        return DBToolsRxUtil.just(new Func0<List<I>>() {
-            @Override
-            public List<I> call() {
-                return findAllValuesBySelection(getDatabaseName(), valueType, column, selection, selectionArgs, null);
-            }
-        });
+    public <I> Observable<I> findAllValuesBySelectionRx(@Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs) {
+        return findAllValuesBySelectionRx(getDatabaseName(), valueType, column, selection, selectionArgs, null);
     }
 
     /**
@@ -734,16 +671,11 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @param selection     Query selection
      * @param selectionArgs Query parameters
      * @param orderBy       Query order by
-     * @return query results List or empty List returned
+     * @return Observable of all values
      */
     @Nonnull
-    public <I> Observable<List<I>> findAllValuesBySelectionRx(@Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
-        return DBToolsRxUtil.just(new Func0<List<I>>() {
-            @Override
-            public List<I> call() {
-                return findAllValuesBySelection(getDatabaseName(), valueType, column, selection, selectionArgs, orderBy);
-            }
-        });
+    public <I> Observable<I> findAllValuesBySelectionRx(@Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
+        return findAllValuesBySelectionRx(getDatabaseName(), valueType, column, selection, selectionArgs, orderBy);
     }
 
     /**
@@ -755,16 +687,11 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @param selection     Query selection
      * @param selectionArgs Query parameters
      * @param orderBy       Query order by
-     * @return query results List or empty List returned
+     * @return Observable of all values
      */
     @Nonnull
-    public <I> Observable<List<I>> findAllValuesBySelectionRx(@Nonnull final String databaseName, @Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
-        return DBToolsRxUtil.just(new Func0<List<I>>() {
-            @Override
-            public List<I> call() {
-                return findAllValuesBySelection(getReadableDatabase(databaseName), getTableName(), valueType, column, selection, selectionArgs, orderBy);
-            }
-        });
+    public <I> Observable<I> findAllValuesBySelectionRx(@Nonnull final String databaseName, @Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
+        return findAllValuesBySelectionRx(getReadableDatabase(databaseName), getTableName(), valueType, column, selection, selectionArgs, orderBy);
     }
 
     /**
@@ -777,17 +704,29 @@ public abstract class RxAndroidBaseManager<T extends AndroidBaseRecord> extends 
      * @param selection     Query selection
      * @param selectionArgs Query parameters
      * @param orderBy       Query order by
-     * @return query results List or empty List returned
+     * @return Observable of all values
      */
     @Nonnull
-    public static <I> Observable<List<I>> findAllValuesBySelectionRx(@Nonnull final DatabaseWrapper database, @Nonnull final String tableName, @Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
-        return DBToolsRxUtil.just(new Func0<List<I>>() {
+    public static <I> Observable<I> findAllValuesBySelectionRx(@Nonnull final DatabaseWrapper database, @Nonnull final String tableName, @Nonnull final Class<I> valueType, @Nonnull final String column, @Nullable final String selection, @Nullable final String[] selectionArgs, @Nullable final String orderBy) {
+        return Observable.create(new Observable.OnSubscribe<I>() {
             @Override
-            public List<I> call() {
-                return findAllValuesBySelection(database, tableName, valueType, column, selection, selectionArgs, orderBy);
+            public void call(Subscriber<? super I> subscriber) {
+                DatabaseValue<I> databaseValue = getDatabaseValue(valueType);
+
+                Cursor c = database.query(tableName, new String[]{column}, selection, selectionArgs, null, null, orderBy);
+                if (c != null) {
+                    if (c.moveToFirst()) {
+                        do {
+                            subscriber.onNext(databaseValue.getColumnValue(c, 0, null));
+                        } while (c.moveToNext());
+                    }
+                    c.close();
+                    subscriber.onCompleted();
+                }
             }
         });
     }
+
 
     private static <I> Observable<DatabaseValue<I>> getDatabaseValueRx(final Class<I> type) {
         return DBToolsRxUtil.just(new Func0<DatabaseValue<I>>() {