+++ src/main/java/org/dbtools/android/domain/event/DatabaseChangeType.java
@@ -1,0 +1,5 @@
+package org.dbtools.android.domain.event;
+
+public enum DatabaseChangeType {
+    INSERT, UPDATE, DELETE
+}